refactored annotation and made models findable for migrations
